Adding in Child Classes for calculations
