+++ src/main/java/com/lmig/gfc/rpn/models/Divider.java
@@ -2,15 +2,15 @@ package com.lmig.gfc.rpn.models;
 
 import java.util.Stack;
 
-public class Subtractor extends TwoNumberCalculation implements Undoer {
+public class Divider extends TwoNumberCalculation implements Undoer {
 
-	public Subtractor(Stack<Double> stack) {
+	public Divider(Stack<Double> stack) {
 		super(stack);
 	}
 	
 	@Override
 	protected double doMath(double firstNumber, double secondNumber) {
-		return (secondNumber - firstNumber);
+		return (firstNumber + secondNumber);
 	}
-	
+
 }
